Add a test case data file with no properties set
